+++ certificate/letsencrypt/dns/netnod/provider.go
@@ -1,0 +1,52 @@
+package netnod
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v4/challenge"
+	"github.com/go-acme/lego/v4/providers/dns/netnod"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+const tokenFieldID = "netnodToken"
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "NETNOD"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsNetnodName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          tokenFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsNetnodToken),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	token, _ := parameters[tokenFieldID].(string)
+
+	cfg := netnod.NewDefaultConfig()
+	cfg.Token = token
+	cfg.TTL = dns.TTL
+	cfg.PropagationTimeout = dns.PropagationTimeout
+	cfg.PollingInterval = dns.PollingInterval
+
+	return netnod.NewDNSProviderConfig(cfg)
+}